Renaming classes for consistency across the project.
